Use main.py for starting and stopping the bot
